Doc-Q&A-bot/requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
faiss-cpu==1.8.0
pypdf==4.3.1
python-docx==1.1.2
sentence-transformers==3.0.1
python-dotenv==1.0.1
streamlit==1.37.0
tiktoken==0.7.0
docx2txt==0.8